+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file. This change
 - `rand-dna` doesn’t take an `elems` vector anymore to select from, which could have arbitrary elements. Its primary function is now better served by the `-weighted` variant and this ensures that it always returns a valid formDNA.
 - `expr/mark-exprs` is deprecated to be replaced by `expr/form-marked` and `expr/make-marked` (the “unmarked” variant), to avoid the need for an options map.
 - `expr/nest-exprs` is deprecated to be replaced by `expr/form-nested-l`, `expr/form-nested-r` (l/r distinguishes the nesting direction) and `expr/make-nested-l`, `expr/make-nested-r` (the “unmarked” variants), again to avoid the need for an options map.
+- `expr/simplify-expr-chain` is deprecated to be replaced by `expr/simplify-nested-l` and `expr/simplify-nested-r` (l/r distinguishes, again, the nesting direction). This is in line with the change above.
 
 ### Fixed